pyproject.toml
[project]
name = "erplora-hub"
version = "0.12.1"
description = "ERPlora Hub - Business Management Desktop Application with Django and PyInstaller"
readme = "docs/README.md"
requires-python = ">=3.11"
license = { text = "BUSL-1.1" }
authors = [
    { name = "ERPlora Team", email = "dev@erplora.com" }
]
keywords = ["erp", "business-management", "pos", "django", "desktop", "pyinstaller"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Django :: 5.2",
    "Intended Audience :: End Users/Desktop",
    "License :: Other/Proprietary License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # === CORE HUB ===
    "Django>=5.2.7",
    "python-decouple>=3.8",
    "gunicorn>=21.0.0", # WSGI server para Docker (web deployment)
    "pywebview>=5.0.0", # Desktop webview (desktop deployment)
    "pyinstaller>=6.0.0", # Desktop builds (desktop deployment)
    # === PAQUETES DJANGO COMUNES ===
    "django-components>=0.123", # Componentes reutilizables
    "django-filter>=24.0", # Filtros avanzados para querysets
    "djangorestframework>=3.15.0", # Django REST Framework
    "django-cors-headers>=4.0.0", # CORS para APIs
    "django-extensions>=3.2.0", # Comandos útiles (shell_plus, etc)
    "django-money>=3.0.0", # Currency handling
    "django-htmx>=1.17.0", # HTMX integration for Django
    "django-health-check>=3.18.0", # Health check endpoint for monitoring
    # === LIBRERÍAS PARA MODULES (25 pre-empaquetadas) ===
    # Ver: config/module_allowed_deps.py para casos de uso
    # Imágenes & Media
    "Pillow>=10.0.0",
    "qrcode>=7.4.0",
    "python-barcode>=0.15.0",
    # Office & Reports
    "openpyxl>=3.1.0",
    "reportlab>=4.0.0",
    "PyPDF2>=3.0.0",
    # Impresión
    "python-escpos>=3.0",
    # XML & Facturación Electrónica
    "lxml>=5.0.0",
    "xmltodict>=0.13.0",
    "signxml>=3.2.0",
    "cryptography>=42.0.0",
    "zeep>=4.2.0",
    # Network & Auth
    "requests>=2.31.0",
    "websockets>=12.0",
    "websocket-client>=1.7.0", # Sync WebSocket client for Hub-to-Cloud
    "PyJWT>=2.8.0",
    # Fechas & Localización
    "python-dateutil>=2.8.2",
    "pytz>=2024.1",
    "phonenumbers>=8.13.0",
    # Pagos
    "stripe>=7.0.0",
    # Data & Analysis
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    # Hardware POS
    "pyserial>=3.5", # Puerto serial (RS232, cajón de dinero)
    "pyusb>=1.2.1", # USB devices (impresoras, scanners)
    "python-escpos>=3.0", # Impresoras térmicas ESC/POS (USB/LAN/Serial)
    "evdev>=1.6.0; sys_platform == 'linux'", # Barcode scanners (Linux)
    "pywinusb>=0.4.2; sys_platform == 'win32'", # USB HID devices (Windows)
    # Email & Validation
    "email-validator>=2.1.0",
    # Utils
    "python-slugify>=8.0.0",
    "pydantic>=2.5.0",
    # Markdown rendering
    "markdown>=3.5.0",
    # Scraping
    "beautifulsoup4>=4.12.0",
    "whitenoise>=6.11.0",
    "django-debug-toolbar>=6.1.0",
    # Background scheduler for automated backups
    "APScheduler>=3.10.0",
    # S3-compatible storage for media files (Hetzner Object Storage)
    "boto3>=1.34.0",
    "django-storages>=1.14.0",
    "drf-spectacular[sidecar]>=0.29.0",
    # PostgreSQL support (for Hub plan with shared database)
    "dj-database-url>=2.3.0",
    "psycopg2-binary>=2.9.10",
]

[project.optional-dependencies]
dev = [
    "coverage>=7.11.0",
    "Faker>=37.12.0",
    "freezegun>=1.4.0",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.11.1",
    "python-semantic-release>=9.0.0",
    "responses>=0.25.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["apps", "config", "."]
only-packages = false

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
upload_to_vcs_release = true
build_command = ""

[tool.semantic_release.branches.main]
match = "main"
prerelease = false

[tool.semantic_release.branches.staging]
match = "staging"
prerelease = true
prerelease_token = "rc"

[tool.semantic_release.changelog]
changelog_file = "CHANGELOG.md"
exclude_commit_patterns = [
    "^Merge",
    "^ci:",
    "^docs:",
    "^chore\\(release\\):",
]

[tool.semantic_release.commit_parser_options]
allowed_tags = [
    "feat",
    "fix",
    "perf",
    "refactor",
]
minor_tags = ["feat"]
patch_tags = ["fix", "perf", "refactor"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
testpaths = ["tests"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "pos: POS tests",
    "modules: Module system tests",
    "hardware: Hardware integration tests",
    "slow: Slow running tests",
    "playwright: Playwright E2E tests",
    "e2e: End-to-end tests",
]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=apps",
    "--cov-report=html",
    "--cov-report=term-missing:skip-covered",
]

[tool.coverage.run]
source = ["apps"]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/test_*.py",
    "*/__init__.py",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = true
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
profile = "black"
line_length = 88
skip_glob = ["*/migrations/*"]
known_django = ["django"]
known_first_party = ["apps", "config"]
sections = ["FUTURE", "STDLIB", "DJANGO", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = ["mypy_django_plugin.main"]

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings"

[tool.ruff]
line-length = 88
target-version = "py311"
extend-exclude = [
    "migrations",
    "*.pyc",
    ".git",
    ".mypy_cache",
    ".pytest_cache",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pyinstaller]
name = "ERPlora Hub"
icon = "assets/icon.ico"

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-django>=4.11.1",
]
